backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
motor>=3.3.2
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
groq>=0.4.2
pytesseract>=0.3.10
Pillow>=10.0.0
pdfplumber>=0.10.4
# pymupdf often causes build issues on windows without generic wheels, using optional install logic in code if missing or try standard
pymupdf>=1.23.0
pymongo>=4.6.0
python-dotenv>=1.0.1
httpx>=0.26.0
email-validator>=2.1.0
jinja2>=3.1.3
